sua giaodien

+++ app/src/main/java/com/example/appquanlynhansu/Model/Luong.kt
@@ -5,7 +5,7 @@ import org.bson.types.ObjectId
 data class Luong(
     val nhanvienId: ObjectId,
     val thang: Double,
-    val luongcoban: Double,
-    val thuong: Double,
-    val tongluong:Double,
+    val luongcoban: Long,
+    val thuong: Long,
+    val tongluong:Long,
 )
